backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
requests>=2.32.0
pandas>=2.2.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
websockets>=12.0
pydantic-settings>=2.0.0

